Refactor abstract base classes for sessions
