+++ src/Ydb.Sdk/src/Services/Table/TableClient.cs
@@ -1,5 +1,5 @@
 ﻿using Ydb.Sdk.Client;
-using Ydb.Sdk.Services.Shared;
+using Ydb.Sdk.Services.Sessions;
 
 namespace Ydb.Sdk.Services.Table;
